Add files via upload
This was my submission for my INFO 303 Final Project. INFO 303 introduced me to the basics of Machine Learning algorithms in Python, the class also increased my aptitude using Pandas, Numpy, and other python packages. Please read 'Mattson_Nutrition_Project_Directions.pdf' for an overview of the project and what was expected. To view the specific machine learning algorithms, and to find out which one was actually implemented to make predictions, view the 'Algorithm_testing.py' file. Among the other files are a brief synopsis of the meta decisions I made when deciding how to handle and clean the data. I was unable to upload the data used to train the model because the file was far larger than the 25MB restriction. To get a sense of the characteristics of the data and the features used in training and testing, as well as the Support Vector Machine used to trim down the number of features through aggregation, view the 'Data_Prep&Features_standardization.py' file.